requirements.txt
torch==1.9.0
torchvision==0.10.0
PyYAML==6.0.1
easydict==1.10
tensorboardX==2.6.2.2
opencv-python==3.4.2.16
numpy==1.19.5
einops==0.4.1
tabulate==0.8.10
scikit-learn==0.24.2
Pillow==8.4.0